Drying code
Moved some shared concurrency code to a new file
