+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "concurrent_prime_sieve"
-version = "0.2.0"
+version = "0.2.1"
 authors = ["FrogBomb <TLBlanchet@gmail.com>"]
 keywords = ["primes", "concurrent", "fast", "generator", "math"]
