+++ src/filter.rs
@@ -1,6 +1,5 @@
 use std;
-use std::thread;
-use std::sync::mpsc;
+use concurrent_help::to_concurrent_on_section;
 use num_cpus;
 // macro_rules! if_any_divides {
 //      ( $( $div:tt ),* | $n:ident $if_b:block ) => { if $(($n % $div == 0))|* $if_b};
@@ -43,45 +42,7 @@ pub fn prime_filter_concurrently(max_num: usize, threads: usize) -> Vec<bool>{
     prime_filter_section_concurrently(0, max_num, threads)
 }
 pub fn prime_filter_section_concurrently(min_num:usize, max_num: usize, threads: usize) -> Vec<bool>{
-    let mut res_vec: Vec<Vec<bool>> = vec![vec![]; threads];
-    let seg_size = (max_num - min_num)/threads;
-    let (tx, rx) = mpsc::channel();
-    for i in 0..threads{
-        let (tx, min, max) = (tx.clone(), min_num + seg_size*i,
-                                min_num + seg_size*(i+1));
-        thread::spawn( move || {
-            let to_send = match max-min {
-                0 => vec![],
-                _ => prime_filter_section_sequentially(min, max),
-            };
-            tx.send((i, to_send)).unwrap();
-        });
-    }
-    if (min_num + seg_size*threads) != max_num {
-        res_vec.push(vec![]);
-        let (tx, min, max) = (tx.clone(), min_num + seg_size*threads, max_num);
-        thread::spawn( move || {
-            let to_send = match max-min {
-                0 => vec![],
-                _ => prime_filter_section_sequentially(min, max),
-            };
-            tx.send((threads, to_send)).unwrap();
-        });
-
-        let (i, p_sec) = match rx.recv(){
-            Ok(mes) => mes,
-            Err(e) => panic!(e.to_string()),
-        };
-        res_vec[i] = p_sec;
-    };
-    for _ in 0..threads{
-        let (i, p_sec) = match rx.recv(){
-            Ok(mes) => mes,
-            Err(e) => panic!(e.to_string()),
-        };
-        res_vec[i] = p_sec;
-    }
-    res_vec.into_iter().flat_map(|x| x).collect()
+    to_concurrent_on_section(prime_filter_section_sequentially, min_num, max_num, threads)
 }
 
 fn int_sqrt(n:usize) -> usize{
